Update _cli.py
Add formats factory instance that uses the `parsed_args.format` arg to retrieve the formatter class, defaults to JSON if not passed or if given a bad/mistyped name.

Might consider error in the case of a bad name given. Shouldn't be too difficult to add that ability if its wanted.